Renamed tracking folder to scoring folder
