+++ src/pieces/discussion/menu/staff/managePostScoring/AddPostAwardModal.ts
@@ -1,6 +1,6 @@
 import { ActionRowBuilder, ButtonInteraction, Client, Emoji, ModalSubmitInteraction, TextInputBuilder, parseEmoji } from "discord.js";
 import { AWARD_POINTS_INPUT_ID, AWARD_STAFF_ONLY_INPUT_ID, AWARD_UNICODE_INPUT_ID, awardPointsInputActionRow, awardStaffOnlyInputActionRow, awardUnicodeInputActionRow, openPostScoringModal, updateCourse } from "./ScoringModalUtilities";
-import { scoreAllThreads } from "../../../../../pieces/discussion/tracking/scoreFunctions";
+import { scoreAllThreads } from "../../../scoring/scoreFunctions";
 import { getCourseByName } from "../../../../../generalUtilities/getCourseByName";
 import { AwardSpecs } from "src/generalModels/DiscussionScoring";
