+++ src/pieces/discussion/scoring/scoreFunctions.ts
